extracted createLightspeedSale
